# aspnetcore-server-generated/src/IO.Swagger/Controllers/CapabilitiesApi.cs
/*
 * Service-Oriented Vehicle Diagnostic API
 *
 * Definition of ASAM SOVD API
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class CapabilitiesApiController : ControllerBase
    { 
        /// <summary>
        /// 
        /// </summary>
        /// <remarks>For entities, resources, and resource collections, a corresponding online capability description can be requested at the SOVD API. The resulting online capability description is a self-contained, valid OpenAPI specification. The description contains information which refers to the creation, reading, updating or deleting of the respective element and its direct child elements as defined by the SOVD standard. An online capability description is requested for a data resource /{entity-path}/data/{data-id} via GET /{entity-path}/data/{data-id}/docs. For example, to enrich the responses with additional details such as the underlying schema of the response using the “include-schema” query parameter on the request. Moreover, for each supported HTTP method all defined HTTP response status codes are documented together with the data model of the resulting response payloads.</remarks>
        /// <param name="anyPath">Path to entity, resource or resource collection for which to request the online capability description.</param>
        /// <response code="200">The online capability description of the underlying resource in form of an OpenAPI specification in JSON document format.</response>
        /// <response code="0">An unexpected request occurred</response>
        [HttpGet]
        [Route("/{any-path}/docs")]
        [ValidateModelState]
        [SwaggerOperation("AnyPathDocsGet")]
        public virtual IActionResult AnyPathDocsGet([FromRoute][Required]Object anyPath)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            return StatusCode(200);

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);


            throw new NotImplementedException();
        }
    }
}