# aspnetcore-server-generated/src/IO.Swagger/Controllers/ConfigurationsApi.cs
/*
 * Service-Oriented Vehicle Diagnostic API
 *
 * Definition of ASAM SOVD API
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class ConfigurationsApiController : ControllerBase
    { 
        /// <summary>
        /// 
        /// </summary>
        /// <remarks>This method is used to read a configuration from an entity, which can either be * bulk data or * parameter data (MIME type: application/json) The SOVD standard follows the server-driven content negotiation where the client defines its preferences using the Accept request header. If no Accept request header is provided, the SOVD server returns a default representation.</remarks>
        /// <param name="entityCollection">Path to an entity collection</param>
        /// <param name="entityId">ID of an entity</param>
        /// <param name="configurationId">The identifier for the configuration</param>
        /// <param name="includeSchema">Specifies whether the response should include schema information or not. Default: false</param>
        /// <response code="200">Response: Read Configurations</response>
        /// <response code="406">bulk data: The configuration cannot be provided in the requested MIME type. parameter data: The configuration cannot be provided as parameter data, i.e., the configuration cannot be provided as JSON.</response>
        /// <response code="0">An unexpected request occurred</response>
        [HttpGet]
        [Route("/{entity-collection}/{entity-id}/configurations/{configuration-id}")]
        [ValidateModelState]
        [SwaggerOperation("EntityCollectionEntityIdConfigurationsConfigurationIdGet")]
        public virtual IActionResult EntityCollectionEntityIdConfigurationsConfigurationIdGet([FromRoute][Required]Object entityCollection, [FromRoute][Required]Object entityId, [FromRoute][Required]Object configurationId, [FromQuery]Object includeSchema)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            return StatusCode(200);

            //TODO: Uncomment the next line to return response 406 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(406);

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);


            throw new NotImplementedException();
        }

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>This method is used to write a configuration to an entity, which can either be • bulk data, or • parameter data (MIME type: application/json) The SOVD client shall tell the SOVD server in the Content-Type header, which MIME type is written to it.</remarks>
        /// <param name="entityCollection">Path to an entity collection</param>
        /// <param name="entityId">ID of an entity</param>
        /// <param name="configurationId">The identifier for the configuration</param>
        /// <response code="204">Configuration written.</response>
        /// <response code="400">The request body does either not contain all required parameter values or the signature for the provided values is wrong.</response>
        /// <response code="409">An unexpected request occurred</response>
        /// <response code="0">An unexpected request occurred</response>
        [HttpPut]
        [Route("/{entity-collection}/{entity-id}/configurations/{configuration-id}")]
        [ValidateModelState]
        [SwaggerOperation("EntityCollectionEntityIdConfigurationsConfigurationIdPut")]
        public virtual IActionResult EntityCollectionEntityIdConfigurationsConfigurationIdPut([FromRoute][Required]Object entityCollection, [FromRoute][Required]Object entityId, [FromRoute][Required]Object configurationId)
        { 
            //TODO: Uncomment the next line to return response 204 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            return StatusCode(204);

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            //TODO: Uncomment the next line to return response 409 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(409);

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);


            throw new NotImplementedException();
        }

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>Provides the configurations provided by an entity.</remarks>
        /// <param name="entityCollection">Path to an entity collection</param>
        /// <param name="entityId">ID of an entity</param>
        /// <param name="includeSchema">Specifies whether the response should include schema information or not. Default: false</param>
        /// <response code="200">Response: Retrieve list of all configurations provided by the entity.</response>
        /// <response code="0">An unexpected request occurred</response>
        [HttpGet]
        [Route("/{entity-collection}/{entity-id}/configurations")]
        [ValidateModelState]
        [SwaggerOperation("EntityCollectionEntityIdConfigurationsGet")]
        public virtual IActionResult EntityCollectionEntityIdConfigurationsGet([FromRoute][Required]Object entityCollection, [FromRoute][Required]Object entityId, [FromQuery]Object includeSchema)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            return StatusCode(200);

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);


            throw new NotImplementedException();
        }
    }
}