# aspnetcore-server-generated/src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Definition of ASAM SOVD API",
    "version" : "1.0.0",
    "title" : "Service-Oriented Vehicle Diagnostic API",
    "contact" : {
      "url" : "https://www.asam.net/project-detail/sovd-service-oriented-vehicle-diagnostics/"
    }
  },
...
  "paths" : {
...
    "/{any-path}/docs" : {
      "get" : {
        "tags" : [ "capabilities" ],
        "description" : "For entities, resources, and resource collections, a corresponding online capability description can be requested at the SOVD API. The resulting online capability description is a self-contained, valid OpenAPI specification. The description contains information which refers to the creation, reading, updating or deleting of the respective element and its direct child elements as defined by the SOVD standard. An online capability description is requested for a data resource /{entity-path}/data/{data-id} via GET /{entity-path}/data/{data-id}/docs. For example, to enrich the responses with additional details such as the underlying schema of the response using the “include-schema” query parameter on the request. Moreover, for each supported HTTP method all defined HTTP response status codes are documented together with the data model of the resulting response payloads.",
        "parameters" : [ {
          "name" : "any-path",
          "in" : "path",
          "description" : "Path to entity, resource or resource collection for which to request the online capability description.",
          "required" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "The online capability description of the underlying resource in form of an OpenAPI specification in JSON document format."
          },
          "default" : {
            "description" : "An unexpected request occurred"
          }
        }
      }
    },
    "/{entity-collection}/{entity-id}/configurations" : {
      "get" : {
        "tags" : [ "configurations" ],
        "description" : "Provides the configurations provided by an entity.",
        "parameters" : [ {
          "name" : "entity-collection",
          "in" : "path",
          "description" : "Path to an entity collection",
          "required" : true
        }, {
          "name" : "entity-id",
          "in" : "path",
          "description" : "ID of an entity",
          "required" : true
        }, {
          "name" : "include-schema",
          "in" : "query",
          "description" : "Specifies whether the response should include schema information or not. Default: false",
          "required" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "Response: Retrieve list of all configurations provided by the entity."
          },
          "default" : {
            "description" : "An unexpected request occurred"
          }
        }
      }
    },
    "/{entity-collection}/{entity-id}/configurations/{configuration-id}" : {
      "get" : {
        "tags" : [ "configurations" ],
        "description" : "This method is used to read a configuration from an entity, which can either be * bulk data or * parameter data (MIME type: application/json) The SOVD standard follows the server-driven content negotiation where the client defines its preferences using the Accept request header. If no Accept request header is provided, the SOVD server returns a default representation.",
        "parameters" : [ {
          "name" : "entity-collection",
          "in" : "path",
          "description" : "Path to an entity collection",
          "required" : true
        }, {
          "name" : "entity-id",
          "in" : "path",
          "description" : "ID of an entity",
          "required" : true
        }, {
          "name" : "configuration-id",
          "in" : "path",
          "description" : "The identifier for the configuration",
          "required" : true
        }, {
          "name" : "include-schema",
          "in" : "query",
          "description" : "Specifies whether the response should include schema information or not. Default: false",
          "required" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "Response: Read Configurations"
          },
          "406" : {
            "description" : "bulk data: The configuration cannot be provided in the requested MIME type. parameter data: The configuration cannot be provided as parameter data, i.e., the configuration cannot be provided as JSON."
          },
          "default" : {
            "description" : "An unexpected request occurred"
          }
        }
      },
      "put" : {
        "tags" : [ "configurations" ],
        "description" : "This method is used to write a configuration to an entity, which can either be • bulk data, or • parameter data (MIME type: application/json) The SOVD client shall tell the SOVD server in the Content-Type header, which MIME type is written to it.",
        "parameters" : [ {
          "name" : "entity-collection",
          "in" : "path",
          "description" : "Path to an entity collection",
          "required" : true
        }, {
          "name" : "entity-id",
          "in" : "path",
          "description" : "ID of an entity",
          "required" : true
        }, {
          "name" : "configuration-id",
          "in" : "path",
          "description" : "The identifier for the configuration",
          "required" : true
        } ],
        "responses" : {
          "204" : {
            "description" : "Configuration written."
          },
          "400" : {
            "description" : "The request body does either not contain all required parameter values or the signature for the provided values is wrong."
          },
          "409" : {
            "description" : "An unexpected request occurred"
          },
          "default" : {
            "description" : "An unexpected request occurred"
          }
        }
      }
    },
...
  },
...
}